please lint.
git-svn-id: file:///svn/unbound/trunk@1433 be551aaa-1e26-0410-a405-d3ace91eadb9

+++ iterator/iterator.c
@@ -244,7 +244,7 @@ error_response_cache(struct module_qstate* qstate, int id, int rcode)
 	/* store in cache */
 	struct reply_info err;
 	memset(&err, 0, sizeof(err));
-	err.flags = BIT_QR | BIT_RA;
+	err.flags = (uint16_t)(BIT_QR | BIT_RA);
 	FLAGS_SET_RCODE(err.flags, rcode);
 	err.qdcount = 1;
 	err.ttl = NORR_TTL;
